mqtt-kafka-bridge/requirements.txt
paho-mqtt==2.0
kafka-python-ng==2.1
